Extract class to test for proposal submissions.
The tests should revolve around submitting a proposal and their primary goal
should be things that test a proposals submit transition.
